docker/requirements.txt
# UGC Face Swapper V1 MVP - Python Dependencies

# Runpod serverless
runpod>=1.3.0

# Cloud storage
boto3==1.34.0

# HTTP requests
requests==2.31.0

# Deep learning
torch==2.1.0
torchvision==0.16.0
--extra-index-url https://download.pytorch.org/whl/cu118

# Computer vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0

# Face detection
retinaface-pytorch==0.0.7
insightface==0.7.3

# Video processing
ffmpeg-python==0.2.0
moviepy==1.0.3

# Utilities
numpy==1.26.3
PyYAML==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

# Logging
python-json-logger==2.0.7

# Wan2.2-Animate dependencies
huggingface-hub[cli]>=0.20.0
einops>=0.7.0
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
omegaconf>=2.3.0
safetensors>=0.4.0
